chore(monorepo): fix inconsistent composer.json names
The psr7 error handler, the pimple-bridge and the
illuminate container bridge had a typo in their
composer.json files.

This also caused the issues when trying to install
the monorepo which is why we initially removed
those specifc names via the monorepo-builder
config.

This now adds consistent names to all packages.