ci(monorepo): enable the monorepo split workflow
Use a custom script in bin/php to split the
monorepo.

The symplify action had no error-handling
at all which led to many releases that
failed but were reported as successful.

GPG signing was also not possible
since the action run in its own
docker container.

Furthermore, it was not possible to
push to a branch dynamically, master
was required.

Our custom php bin script handles
all of these scenarios now and
gives as more control in the
future if we want to decide
to only split the repo on minor
versions.