fix(testing-bundle): fix request generation of test browser
The Browser class had some missing pieces in the
code that transforms the Symfony Browserkit response
into a PSR7 response.

In particular, HTTP_XXX and CONTENT_XXX server params
were not transformed to headers in the PSR7 request.

Furthermore, request were marked as API requests
if the API prefix in the configuration was "/".

Fix: #140
Fix: #141