fix(http-routing-bundle): enforce request data is not slashed
WordPress will still completely nuke the request variables
by slashing them with wp_magic_quotes().

There is an open ticket for this since 10+ years, but
we suppose it will never get fixed.

With the http-routing-bundle we are always getting the real
server variables, but this change ensures and enforces this.

See: https://core.trac.wordpress.org/ticket/18322