fix(better-wp-cli-testing): use temp stream instead of memory as test stdin
It should have been a tmp stream from the beginning. They are identical with the
only meaningful difference being that tmp will work with stream_select (always returns 1) while memory will throw a fatal error with stream_select on PHP8+.

This is an issue because stream_select is the recommended way to check
if reading from STDIN will block forever because it's interactive/the terminal.