fix(kernel): don't check if directories are readable in constructor
BREAKING CHANGE: Previously creating an instance of the kernel class always asserted
that cache/log/config dirs are
readable.
This created a lot of development complexities as these directories
always had to be pre-created before a kernel instance could be created.
It also creates runtime overhead for little additional safety.
There is no way to ensure that a directory did not go missing between
Directories::__construct(), and the usage of any directory.
Callers of the Directory class are now responsible for error handling,
as they should always have.